requirements.txt
flask
flask-smorest
python-dotenv
flask-sqlalchemy
flask-jwt-extended
flask-migrate
passlib